VoteService와 VoteServiceImpl로 분리
