+++ src/main/java/com/valanse/valanse/service/VoteService/VoteService.java
@@ -1,61 +1,13 @@
 package com.valanse.valanse.service.VoteService;
 
-import com.valanse.valanse.common.api.ApiException;
-import com.valanse.valanse.domain.Member;
-import com.valanse.valanse.domain.Vote;
-import com.valanse.valanse.dto.Vote.VoteResponseDto;
-import com.valanse.valanse.repository.MemberRepository;
-import com.valanse.valanse.repository.VoteRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.stereotype.Service;
 import com.valanse.valanse.domain.enums.VoteCategory;
+import com.valanse.valanse.dto.Vote.VoteResponseDto;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
-@Service
-@RequiredArgsConstructor
-public class VoteService {
+public interface VoteService {
 
-    private final VoteRepository voteRepository;
-    private final MemberRepository memberRepository;
+    List<VoteResponseDto> getMyCreatedVotes(Long memberId, String sort, VoteCategory category);
 
-    public List<VoteResponseDto> getMyCreatedVotes(Long memberId, String sort, VoteCategory category) {
-        Member member = memberRepository.findByIdAndDeletedAtIsNull(memberId)
-                .orElseThrow(() -> new ApiException("회원이 존재하지 않습니다.", HttpStatus.NOT_FOUND));
-
-        List<Vote> votes;
-        if (category != null) {
-            String categoryName = category.name();
-            votes = sort.equals("latest") ?
-                    voteRepository.findAllByMemberAndCategoryOrderByCreatedAtDesc(member, category) :
-                    voteRepository.findAllByMemberAndCategoryOrderByCreatedAtAsc(member, category);
-        } else {
-            votes = sort.equals("latest") ?
-                    voteRepository.findAllByMemberOrderByCreatedAtDesc(member) :
-                    voteRepository.findAllByMemberOrderByCreatedAtAsc(member);
-        }
-
-        return votes.stream().map(VoteResponseDto::new).collect(Collectors.toList());
-    }
-
-    public List<VoteResponseDto> getMyVotedVotes(Long memberId, String sort, VoteCategory category) {
-        Member member = memberRepository.findByIdAndDeletedAtIsNull(memberId)
-                .orElseThrow(() -> new ApiException("회원이 존재하지 않습니다.", HttpStatus.NOT_FOUND));
-
-        List<Vote> votes;
-        if (category != null) {
-            String categoryName = category.name();
-            votes = sort.equals("latest") ?
-                    voteRepository.findAllByMemberVotedAndCategoryOrderByCreatedAtDesc(member, category) :
-                    voteRepository.findAllByMemberVotedAndCategoryOrderByCreatedAtAsc(member, category);
-        } else {
-                votes = sort.equals("latest") ?
-                        voteRepository.findAllByMemberVotedOrderByCreatedAtDesc(member) :
-                        voteRepository.findAllByMemberVotedOrderByCreatedAtAsc(member);
-        }
-
-        return votes.stream().map(VoteResponseDto::new).collect(Collectors.toList());
-    }
+    List<VoteResponseDto> getMyVotedVotes(Long memberId, String sort, VoteCategory category);
 }
